Fix CI env helper visibility
